+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
  
 ## [Unreleased]
+#### Fixed
+- Using correct const for EntitlementPrefix in PerunEntitlement.php
 
 ## [v3.7.0]
 #### Changed